+++ libtelemetry/src/full/java/com/mapbox/android/telemetry/MapboxTelemetry.java
@@ -18,6 +18,7 @@ import java.util.concurrent.ThreadFactory;
 import java.util.concurrent.LinkedBlockingQueue;
 import java.util.concurrent.RejectedExecutionException;
 import java.util.concurrent.atomic.AtomicReference;
+import java.util.regex.Pattern;
 
 import android.support.annotation.NonNull;
 import android.support.v4.content.LocalBroadcastManager;
@@ -475,5 +476,19 @@ public class MapboxTelemetry implements FullQueueCallback, ServiceTaskCallback {
         }
       };
     }
+  }
+
+  @SuppressWarnings("WeakerAccess")
+  public synchronized boolean setBaseUrl(String eventsHost) {
+    if (isValidUrl(eventsHost) && checkNetworkAndParameters()) {
+      telemetryClient.setBaseUrl(eventsHost);
+      return true;
+    }
+    return false;
+  }
+
+  private static boolean isValidUrl(String eventsHost) {
+    Pattern urlPattern = Pattern.compile("^[a-z0-9]+([\\-.][a-z0-9]+)*\\.[a-z]{2,5}(:[0-9]{1,5})?(/.*)?$");
+    return eventsHost != null && !eventsHost.isEmpty() && urlPattern.matcher(eventsHost).matches();
   }
 }