+++ CHANGELOG.md
@@ -4,6 +4,13 @@ Mapbox welcomes participation and contributions from everyone.
 
 ## Mapbox Android Telemetry
 
+## v4.6.1
+- Fix crash `setBaseUrl()` on null `TelemetryClient` [#423](https://github.com/mapbox/mapbox-events-android/pull/423)
+
+## v4.6.0
+- Add public api to change base url for telemetry endpoint [#420](https://github.com/mapbox/mapbox-events-android/pull/420)
+- Telemetry metrics [#397](https://github.com/mapbox/mapbox-events-android/pull/397)
+
 ### v4.5.1
 - Revert okhttp to version 3.12.0 [#413](https://github.com/mapbox/mapbox-events-android/pull/413)